✨ : 빠른챔피언 클릭시 기능
